requirements.txt
pdfplumber
pydantic
python-dotenv
pdfplumber
